requirements.txt
google-adk
litellm
ag_ui_adk
uvicorn 
fastapi